requirements.txt
fire
python-dotenv
setuptools
yfinance>=0.2.54
gradio
llama-stack-client>=0.1.0
##Claude
PyPDF2
python-docx
mammoth
requests
asyncio
##DeepSeek
llama-index
python-dotenv
langchain
sentence-transformers
faiss-cpu
##juypter notebook
ipykernel
## simple frontend
gradio
